refactor(timeline): move long-lived tasks to a new `tasks` mod
Only code motion.
